+++ deps.ts
@@ -1,0 +1,3 @@
+export { Application, send } from 'https://deno.land/x/oak/mod.ts'
+export { Router } from 'https://deno.land/x/oak/mod.ts'
+import "https://deno.land/x/dotenv/load.ts";
