+++ srv/routes.ts
@@ -1,4 +1,4 @@
-import { Router } from 'https://deno.land/x/oak/mod.ts'
+import { Router } from '../deps.ts';
 import { getConfig, isAllowed, getDBPath } from './controller.ts'
 
 const router = new Router()
